♻️ refactor: BookmarkAccordian 관련 파일 위치 수정 [#25]
Explained:
- features 에서 pages로 이동
