--- src/features/bookmark/boomarksAccordian/index.ts
@@ -1,3 +1,0 @@
-import BookmarksAccordian from '@/features/bookmark/boomarksAccordian/bookmarksAccordian.ui';
-
-export default BookmarksAccordian;
