+++ src/pages/modal/sidebar/ui/bookmarkAccordian/index.ts
@@ -1,0 +1,1 @@
+export { BookmarkAccordian } from './bookmarkAccordian.ui';
